monitoring_api/requirements.txt
fastapi
uvicorn
python-dotenv
python-json-logger
passlib[bcrypt]
psutil
requests
slowapi
colorlog
